Explicitly export icons from main.ts
Now different builds trageting different apps can chose which icons to
include in main.ts instead of in Icons.vue. By not affecting the exports
in Icons.vue, the build process (probably the linter only) will not complain
about missing exports anymore when checking files in the folder structure
that are not actually part of the build but still checked for some reason.