AddressInput: ensure correct space width for tailing whitespace
The word spacing set via css is only applied to spaces that are actually between
words which is not the case for an ending space and the caret after an ending space
therefore gets rendered at the wrong position. To avoid that we add a zero-width
space as an artificial word.