Carousel: automatically calculate min-height
Calculate min-height as height of the largest entry such that the entry
fits into the component when it's in the front.
To calculate the required carousel width, we would need to perspectively
scale the distance between the points where the perspective rays starting
at the user meet the circle (with increased radius accounting for the
entry widths) as tangents. We're not currently doing this.