Tooltip: remove need for measurements initially and while not visible
Previously, a complicated and costly initial measurement was taken and continous
updates during scrolls were calculated even when the tooltip was not visible.
This was to make sure that the tooltip is already correctly positioned on show.
This commit now removes the necessity to do so by disabbling position transitions
on show until the tooltip is correctly positioned.