Revert "storybook: disable class name mangling on build-storybook"
This reverts commit 6680f9c4401d3285f186c154ba6d31135b2e32a6.
Reverting this commit such that the build setup for the storybook demo
page resembles the same class name mangling behavior as the normal build
to be able to catch any issues with it. A workaround for the mangling
issue has been implemented in the mean time by using $options.name instead
of consructor.name in I18nMixin.ts without the need to disable class name
mangling entirely which negatively impacts minification effectiveness.