vue.config.js: make qr scanner worker an external dependency
This avoids webpack generating a separate chunk for the qr scanner worker that
would need to be moved over to the consuming app. Instead, the import and
generation of a new chunk for the qr scanner worker is now handled during the
build of the consuming app without the need for any additional config.