vue.config.js: resolve qr-scanner worker by package instead of by node_modules path
While resolving the external dependency for the qr-scanner worker by absolute path
/node_modules/qr-scanner/qr-scanner-worker.min.js works when building the Wallet, it
did not work for building the gift-cards app. This might be due to an older Webpack
version in use in gift-cards.

Resolving the file by package via qr-scanner/qr-scanner-worker.min.js works for both,
building the Wallet and the gift-cards app.