Tooltip: now allow to place the tooltip in any position
Before the Tooltip was limited at a horizontal position as the first preferred position, and an optional vertical one as the second preferred position. Now every single combination is possible and result in a different position and a single preferred position will result in a centered tooltip. The tooltip will also try to stay within the container bounds as much as possible to stay visible. A background prop was also added to allow to easily customise the tooltip background.