package.json: add script for updating caniuse browser support list
This is used for babel transpilation targets via browserslist and is needed to be updated
sporadically.

We don't follow the official usage instructions to do "npx update-browserslist-db@latest"
to reduce the risk of blindly executing code of a malicious version of the script.
Instead, the package update-browserslist-db should be updated manually occasionally,
which doesn't have to be regularly though.