Update @nimiq/utils for new FiatApi and make provider customizable
CoinGecko is becoming increasingly strict with unauthenticated access to their
API without API key. For example, wallet.nimiq.com has already been blocked from
unauthenticated access, thus the Wallet has been switched to CryptoCompare as
fiat api provider, while hub.nimiq.com and keyguard.nimiq.com are not currently
blocked. As the only vue-component that currently makes use of the FiatApi is
PaymentInfoLine which is used only in the Hub, we make the provider customizable
in PyamentInfoLine, but don't have to switch the default to CryptoCompare yet.